backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Authentication
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.11
SQLAlchemy>=2.0.40

# Google Ads API
google-ads==28.4.1

# Validation
marshmallow==3.20.1

# Environment management
python-dotenv==1.0.0

# Utilities
uuid==1.30

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging
structlog==23.2.0

# Development
black==23.11.0
flake8==6.1.0
